refactor(knex): :art: re-arrange persist classes
